Added Tests for ProductByCode, fixed bug where Product used double instead of BigDecimal

+++ catalog-service/src/main/java/com/maazchowdhry/bookstore/catalog/domain/Product.java
@@ -1,3 +1,5 @@
 package com.maazchowdhry.bookstore.catalog.domain;
 
-public record Product(String code, String name, String description, String imageUrl, double price) {}
+import java.math.BigDecimal;
+
+public record Product(String code, String name, String description, String imageUrl, BigDecimal price) {}
